+++ fs/ntfs.c
@@ -87,7 +87,7 @@ static int apply_fixup(uint8_t *record, uint16_t fixup_offset,
 
 /* Read MFT record by number into mft_buf */
 static int read_mft_record(ntfs_fs_t *fs, uint32_t record_num) {
-    uint32_t lba = fs->mft_lba + record_num * fs->mft_sectors;
+    uint64_t lba = fs->mft_lba + (uint64_t)record_num * fs->mft_sectors;
     if (ata_read_sectors(fs->drive, lba,
                          (uint8_t)fs->mft_sectors, mft_buf) != 0)
         return -1;
@@ -188,8 +188,8 @@ static int read_nonresident(ntfs_fs_t *fs, ntfs_attr_nonresident_t *attr,
             while (remaining > 0) {
                 uint32_t clust_idx    = run_byte_off / fs->cluster_size;
                 uint32_t within_clust = run_byte_off % fs->cluster_size;
-                uint32_t lba = fs->part_lba
-                             + (uint32_t)(lcn + clust_idx) * fs->sectors_per_cluster;
+                uint64_t lba = fs->part_lba
+                             + (uint64_t)(lcn + clust_idx) * fs->sectors_per_cluster;
 
                 if (ata_read_sectors(fs->drive, lba,
                                      fs->sectors_per_cluster, data_buf) != 0)
@@ -441,7 +441,7 @@ vfs_node_t *ntfs_finddir(vfs_node_t *node, const char *name) {
  *  ntfs_read — read bytes from a file
  * ══════════════════════════════════════════════════════════════════════════ */
 
-int ntfs_read(vfs_node_t *node, uint32_t offset, uint32_t size, void *buffer) {
+int ntfs_read(vfs_node_t *node, uint64_t offset, uint32_t size, void *buffer) {
     ntfs_fs_t *fs = (ntfs_fs_t *)node->fs_private;
     if (!fs) return -1;
 
@@ -481,7 +481,7 @@ int ntfs_read(vfs_node_t *node, uint32_t offset, uint32_t size, void *buffer) {
  *  ntfs_mount — mount an NTFS partition (read-only)
  * ══════════════════════════════════════════════════════════════════════════ */
 
-int ntfs_mount(uint8_t drive, uint32_t part_lba, mount_point_t *mp) {
+int ntfs_mount(uint8_t drive, uint64_t part_lba, mount_point_t *mp) {
     uint8_t boot[512];
     if (ata_read_sectors(drive, part_lba, 1, boot) != 0)
         return -1;
@@ -521,7 +521,7 @@ int ntfs_mount(uint8_t drive, uint32_t part_lba, mount_point_t *mp) {
     if (fs->index_record_size == 0) fs->index_record_size = 4096;
 
     /* Absolute LBA of $MFT */
-    fs->mft_lba = part_lba + (uint32_t)bs->mft_cluster * bs->sectors_per_cluster;
+    fs->mft_lba = part_lba + bs->mft_cluster * (uint64_t)bs->sectors_per_cluster;
 
     /* Verify we can read MFT record 0 */
     if (read_mft_record(fs, 0) != 0)
