+++ fs/ntfs.h
@@ -149,20 +149,20 @@ typedef struct __attribute__((packed)) {
 /* ── Internal mounted state ───────────────────────────────────────────── */
 typedef struct {
     uint8_t  drive;
-    uint32_t part_lba;
+    uint64_t part_lba;
     uint8_t  sectors_per_cluster;
     uint32_t cluster_size;
-    uint32_t mft_lba;               /* absolute LBA of $MFT         */
+    uint64_t mft_lba;               /* absolute LBA of $MFT         */
     uint32_t mft_record_size;
     uint32_t mft_sectors;           /* sectors per MFT record        */
     uint32_t index_record_size;
 } ntfs_fs_t;
 
 /* ── Public API ───────────────────────────────────────────────────────── */
-int         ntfs_mount(uint8_t drive, uint32_t part_lba, mount_point_t *mp);
+int         ntfs_mount(uint8_t drive, uint64_t part_lba, mount_point_t *mp);
 dirent_t   *ntfs_readdir(vfs_node_t *node, uint32_t index);
 vfs_node_t *ntfs_finddir(vfs_node_t *node, const char *name);
-int         ntfs_read(vfs_node_t *node, uint32_t offset, uint32_t size,
+int         ntfs_read(vfs_node_t *node, uint64_t offset, uint32_t size,
                       void *buffer);
 
 #endif
