+++ libc/string.h
@@ -2,11 +2,12 @@
 #define STRINGS_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 void int_to_ascii(int n, char str[]);
 void hex_to_ascii(int n, char str[]);
 void reverse(char s[]);
-int strlen(char s[]);
+size_t strlen(char s[]);
 void backspace(char s[]);
 void append(char s[], char n);
 int strcmp(char s1[], char s2[]);
